# Add annotation tracking to Object Detection Confusion Matrices