--- a/src/valor_lite/object_detection/__init__.py
+++ b/src/valor_lite/object_detection/__init__.py
@@ -1,12 +1,4 @@
 from .annotation import Bitmask, BoundingBox, Detection, Polygon
-from .computation import (
-    compute_bbox_iou,
-    compute_bitmask_iou,
-    compute_confusion_matrix,
-    compute_polygon_iou,
-    compute_precion_recall,
-    compute_ranked_pairs,
-)
 from .manager import DataLoader, Evaluator
 from .metric import Metric, MetricType
 
@@ -17,12 +9,6 @@ __all__ = [
     "Polygon",
     "Metric",
     "MetricType",
-    "compute_bbox_iou",
-    "compute_bitmask_iou",
-    "compute_polygon_iou",
-    "compute_ranked_pairs",
-    "compute_precion_recall",
-    "compute_confusion_matrix",
     "DataLoader",
     "Evaluator",
 ]