… benchmarks/benchmark_objdet.py
… examples/testing/torch_detection.ipynb
… src/valor_lite/object_detection/manager.py
… src/valor_lite/object_detection/metric.py
… src/valor_lite/object_detection/utilities.py
… tests/object_detection/conftest.py
… tests/object_detection/test_average_precision.py
… tests/object_detection/test_average_recall.py
… tests/object_detection/test_confusion_matrix.py
… tests/object_detection/test_counts.py
… tests/object_detection/test_dataloader.py
… tests/object_detection/test_evaluator.py
… tests/object_detection/test_f1.py
… tests/object_detection/test_filtering.py
… tests/object_detection/test_pr_curve.py
… tests/object_detection/test_precision.py
… tests/object_detection/test_recall.py
… tests/object_detection/test_schemas.py
… tests/object_detection/test_stability.py