# Changes
- Add `dict` mapping annotation string identifiers to Valor indices
- Add annotation ID's within confusion matrix examples
- Remove limit on confusion matrix examples
- Deprecated Object Detection `DataLoader` see #839 

## New `ConfusionMatric` Format

```
{
    "type": "ConfusionMatrix",
    "value": {
        "confusion_matrix": {
            "v1": {
                "v1": {
                    "count": 1,
                    "examples": [
                        {
                            "datum_id": "uid1",
                            "ground_truth_id": "uid_1_gt_0",
                            "prediction_id": "uid_1_pd_0",
                        }
                    ],
                }
            }
        },
        "unmatched_predictions": {
            "not_v2": {
                "count": 1,
                "examples": [
                    {
                        "datum_id": "uid1",
                        "prediction_id": "uid_1_pd_1",
                    },
                ],
            },
```

# Performance
- Confusion Matrix computation is ~3x slower but now finds **ALL** examples rather than N=3.
- There may be ways to reuse some larger sorts and speed this up but performance seems acceptable for now.
## Before
```
{
    "limit": 5000,
    "n_datums": 5000,
    "n_groundtruths": 36536,
    "n_predictions": 27092,
    "n_labels": 80
    "ingestion": {
        "loading_from_file": "0.51 seconds",
        "numpy_conversion + IoU": "0.52 seconds", <========================
        "ranking_pairs": "0.61 seconds",
        "total": "1.63 seconds"
    },
    "base_evaluation": "0.04 seconds",
    "detailed_evaluation": [
        {
            "n_points": 10,
            "n_examples": 0,
            "computation": "0.42 seconds"
        },
        {
            "n_points": 10,
            "n_examples": 3,
            "computation": "0.66 seconds"  <=========================
        }
    ]
}
```
## After
```
{
    "limit": 5000,
    "n_datums": 4968,
    "n_groundtruths": 36536,
    "n_predictions": 27092,
    "n_labels": 80,
    "ingestion": {
        "loading_from_file": "0.66 seconds",
        "add_data": "1.6 seconds",
        "finalization": "0.25 seconds",
        "total": "2.5 seconds"
    },
    "precision_recall_computation": "0.04 seconds",
    "confusion_matrix_computation": "1.48 seconds"
}
```